+++ src/main/java/com/cells/cells/hyperdensity/compacting/ItemHyperDensityCompactingCellBase.java
@@ -201,7 +201,7 @@ public abstract class ItemHyperDensityCompactingCellBase extends AbstractTieredC
 
     @Override
     public IItemHandler getUpgradesInventory(ItemStack is) {
-        return new CustomCellUpgrades(is, 2, Arrays.asList(
+        return new CustomCellUpgrades(is, CellsConfig.hdCompactingCellUpgradeSlots, Arrays.asList(
             CustomCellUpgrades.CustomUpgrades.OVERFLOW,
             CustomCellUpgrades.CustomUpgrades.OREDICT,
             CustomCellUpgrades.CustomUpgrades.COMPRESSION_TIER,